[*] fix various warnings, prefer uint64_t over ull, better unreachable
- Fixed a bunch of random warnings. There's still a bunch left but
whatever.
- Always prefer uint64_t over unsigned long long. They're the same
  thing, but one is shorter
- Made a platform-agnostic UNREACHABLE() macro so that this can compile
  on MSVC in the future maybe. Also used this macro to mark some
  unenumerated switch/if-paths as unreachable which should result in
  marginally better codegen.

+++ source/gdb_stub.cpp
@@ -115,7 +115,7 @@ static std::optional<CombinedProcessThreadId> ParseProcessThreadId(const std::st
  * @todo we shouldn't be using the HLE::OS::ProcessId type for this protocol-specific quantity!
  */
 static bool IsAnyProcess(HLE::OS::ProcessId pid) {
-    return pid == 0 || pid == -1;
+    return pid == 0 || pid == uint32_t(-1);
 }
 
 /**
@@ -123,7 +123,7 @@ static bool IsAnyProcess(HLE::OS::ProcessId pid) {
  * @todo we shouldn't be using the HLE::OS::ThreadId type for this protocol-specific quantity!
  */
 static bool IsAnyThread(HLE::OS::ThreadId tid) {
-    return tid == 0 || tid == -1;
+    return tid == 0 || tid == uint32_t(-1);
 }
 
 /**
